Move RegistryBuildpack into separate file
* Rename registry.go to registry_cache to more appropriately represent the contents
